solution/pyproject.toml
[project]
name = "solution"
version = "0.1.0"
description = ""
authors = [
    {name = "ShIIIrochka",email = "i@ksenia-travnikova.ru"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "poetry-plugin-export (>=1.9.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "types-pyyaml (>=6.0.12.20241230,<7.0.0.0)",
    "authx (>=1.4.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "icecream (>=2.1.4,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "__init__.py",
    "settings.py",
    "local_settings.py",
    "docs",
    "app_template/*",
    "app/alembic/*",
    ".venv",
    ".vscode",
    "venv",
    "dist",
    "pyproject.toml",
]
line-length = 79
indent-width = 4

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
ruff = "^0.9.2"
pre-commit = "^4.0.1"
pretty-errors = "^1.2.25"
icecream = "^2.1.4"

[tool.poetry.group.tests.dependencies]
tavern = "^2.11.0"

